Still not working
I have tried several ways and times to get the STR working, but none fully fixed it.
Most importantly, it seems like the STR and LDR operations continue into the 4th timestep or T=11. And here, it unexpectedly writes a value to Rx, which i dont want it to.

TO fix it, i tried making sure the two operations do not go to the 4th timestep, but unfortunately, they keep on doings so.

I need to find the and fix bug that does not let them clear after the 3rd timestep